pyproject.toml
[project]
name = "metrics"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Maksim Smirnov", email = "smirnoffmg@gmail.com" }]
dependencies = [
    "numpy>=1.26.3",
    "jira>=3.6.0",
    "python-dateutil>=2.8.2",
    "matplotlib>=3.8.2",
    "pandas>=2.1.4",
    "scipy>=1.11.4",
    "seaborn>=0.13.1",
    "python-decouple>=3.8",
    "pydantic-settings>=2.1.0",
    "click>=8.1.7",
    "dependency-injector>=4.41.0",
]

[dependency-groups]
dev = [
    "ruff>=0.1.11",
    "ipython>=8.19.0",
    "mypy>=1.8.0",
    "pytest>=7.4.4",
    "detect-secrets>=1.4.0",
    "pytest-cov>=4.1.0",
    "types-python-dateutil>=2.9.0.20241003",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["ALL"]
ignore = ["G004", "DTZ001", "S101", "D203", "D213", "COM812"]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["ANN", "D"]

[tool.ruff.format]
# Use Ruff as the formatter
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
plugins = 'numpy.typing.mypy_plugin'

[tool.codespell]
skip = '*.lock'
quiet-level = 3
